requirements.txt
scrapy
bs4
openai
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
selenium
webdriver_manager
mysql-connector-python
gspread
oauth2client
